refactor(api):重构API路由模块并更新依赖- 将`AppState`的引用路径从`crate::AppState`修改为`crate::app_state::AppState` - 更新了多个Rust依赖包及其版本号 - 添加了新的依赖包如`ahash`, `aliasable`, `allocator-api2`等 - 移除了部分旧的依赖配置，优化了项目结构

+++ src/handlers/user.rs
@@ -3,9 +3,9 @@ use axum::{
     response::Json,
     extract::{Path, State},
 };
-use std::sync::{Arc};
+use std::sync::Arc;
 use crate::models::user::{User, CreateUserRequest, UpdateUserRequest};
-use crate::AppState;
+use crate::app_state::AppState;
 use crate::utils::error::AppError;
 
 
@@ -18,7 +18,7 @@ pub async fn get_users(
 
 pub async fn get_user(
     State(state): State<Arc<AppState>>,
-    Path(id): Path<u64>,
+    Path(id): Path<u32>,
 ) -> Result<Json<User>, AppError> {
     let users = state.users.read().map_err(|_| AppError::InternalError)?;
     let user = users.iter().find(|u| u.id == id)
@@ -58,7 +58,7 @@ pub async fn create_user(
 
 pub async fn update_user(
     State(state): State<Arc<AppState>>,
-    Path(id): Path<u64>,
+    Path(id): Path<u32>,
     Json(payload): Json<UpdateUserRequest>,
 ) -> Result<Json<User>, AppError> {
     let mut users = state.users.write().map_err(|_| AppError::InternalError)?;
@@ -85,7 +85,7 @@ pub async fn update_user(
 
 pub async fn delete_user(
     State(state): State<Arc<AppState>>,
-    Path(id): Path<u64>,
+    Path(id): Path<u32>,
 ) -> Result<StatusCode, AppError> {
     let mut users = state.users.write().map_err(|_| AppError::InternalError)?;
     let len_before = users.len();
